Move some functions out of configuration.cpp to program.cpp
They make more sense there.
